Send Deskpro tickets to tech team by default
At the moment we assign all Deskpro tickets to the Notify product team.

In the future, the product team won’t include everyone. This means that
if an technical emergency arises, the developers won’t have visibility
of it.

The technical team need to have visibility of all tickets, so that they
can address ones that need a technical response. They will be able to
triage and re-assign ones that don’t, meaning less work.

This commit changes the default team for support tickets from the
product team (5) to the technical team (6).