Fix spacing of table footers
When implementing the horizontally scrollable tables I messed up the
spacing above the text that says “Only showing the first 50 rows”, where
it’s displayed below a table that doesn’t have horizontal scrolling.

This commit makes the spacing correct when this element is used both
with and without a horizontally scrolling table above it.